% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BalMed_mcmc.R
\name{BalMed_mcmc}
\alias{BalMed_mcmc}
\title{This is the main function of balance mediation analysis
Runs a MCMC chain to sampling all the unknown parameters in the model}
\usage{
BalMed_mcmc(X, Y, Trt, hyperparam, mu0_initial, mu1_initial,
  invSigma_initial, P_initial, z_initial, niters, persaveiters, savepath,
  filename, logfile = "MCMClog.txt")
}
\arguments{
\item{X}{compositional matrix}

\item{Y}{outcome vector}

\item{Trt}{treatment vector currently only accept binary value}

\item{hyperparam}{a list containing all the prior distribution informatiion}

\item{niters}{total number of iterations}

\item{persaveiters}{save chain information for each of the persaveiters iterations}

\item{savepath}{the path to save the current chain info}

\item{filename}{the name of the file to save the current chain info}

\item{logfile}{name of the file to store some log information}

\item{mu0_inital}{starting value for the mean in untreated group}

\item{mu1_inital}{starting value for the mean in the treated group}

\item{invSigma_inital}{starting value for the variance-covariance matrix}

\item{P_inital}{starting value for the latent relative abundance matrix}

\item{z_inital}{starting value for the balance configuration vector z}
}
\description{
This is the main function of balance mediation analysis
Runs a MCMC chain to sampling all the unknown parameters in the model
}
